pyproject.toml
[project]
name = "HoundExpressApi"
version = "0.1.0"
description = "An example Django app running in Docker."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "django==5.2.3",
  "djangorestframework==3.16.0",  
  "celery==5.5.3",
  "django-debug-toolbar==5.2.0",
  "gunicorn==23.0.0",
  "psycopg==3.2.9",
  "redis==6.2.0",
  "ruff==0.12.1",
  "setuptools==80.9.0",
  "whitenoise==6.9.0",
  "requests>=2.32.4",
]

[tool.ruff]
line-length = 79

[tool.ruff.lint]
extend-select = ["I", "SIM"]
